meta-qti-auto-kernel: i2c: Add validation checks and fix inter-frame delay handling
Add a check to reject transfers if the GPII number provided by
the client is outside the valid range (0–14).

Prevent transfers where the TX sub-packet length field is
incorrectly set to a read message in multi random write
scenarios, as this feature is only applicable to write transfers.

Add a check to reject transfers where the TX packet length exceeds
the actual transfer length in multi random write cases.

Correct the bit size of the inter-frame delay field in the config
TRE to 16 bits, as specified in the HPG.

Ensure the oldest callback entry is sent first to the client.
Also, introduce an alignment macro to eliminate structure padding
issues, ensuring accurate callback data is passed to the client.

Additionally, implement logic to convert the received inter-frame
delay value into clock pulses based on the clock frequency,
enabling accurate use of pre and post delay features.

Change-Id: I94406dfef7280f790e08491157210394ea647351
Signed-off-by: Aniket Randive <arandive@qti.qualcomm.com>