meta-qti-auto-kernel: i2c: Update clock cycle formula per latest HPG specification
Update the formula used to calculate clock cycles based on the
latest HPG specification.

The Clock cycle on the bus is now computed as,
  Clock_cycle = Delay(in ns) / (CLK_DIV_VALUE * INPUT_SERIAL_CLOCK_period
(ns) * 64)

This change reflects the updated timing model and ensures accurate
delay configuration for I2C transactions. The maximum supported
delay is 2^20 ns, as defined by the 20-bit width of M_GP_CNT3[19:0].

Change-Id: Id9a7346932ccacf678db2b019c8b9278b0a819c8
Signed-off-by: Aniket Randive <arandive@qti.qualcomm.com>