ccu: Add support for passing error notification event to client
Sometimes the client does not queue any packet on the sync channel,
but if a sync trigger occurs on that channel, the hardware reports
an error event with type I2C_CCU_SYNC_TRE_NOT_FOUND. The client
needs this error event to identify which frame was missed.
This commit adds support to pass such error events to the client.
The driver now sends a poll notification to the client, which
should then read the event using the read API. The event typically
includes the QUP number, GPII number, and error code (0x37).

Change-Id: Ib6fc2a99becd4b7264560ed99c2ce8f7611dafb6
Signed-off-by: Aniket Randive <arandive@qti.qualcomm.com>