meta-qti-auto-kernel: ufs: Skip TX FSM state check for SCMI-based targets
On SCMI-based targets, the PHY is already initialized in GearVM
before the UFS probe begins. As a result, the MPHY TX FSM state
does not follow the usual Disabled to Hibernate transition.
Calling ufs_qcom_check_hibern8() in this context is unnecessary
and may lead to incorrect assumptions about the PHY state.
This change removes the hibern8 state check for SCMI-based targets.

Change-Id: I1544d26444f3de4411dae232565368962d353e98
Signed-off-by: Ram Kumar Dwivedi <rdwivedi@qti.qualcomm.com>