meta-qti-auto-kernel: ufs: Skip TX FSM state check for SCMI-based targets
On SCMI-based targets, the PHY is already initialized in firmware
before the UFS probe begins. As a result, the MPHY TX FSM state
does not follow the usual Disabled to Hibernate transition.
Calling ufs_qcom_check_hibern8() in this context is unnecessary
and may lead to incorrect assumptions about the PHY state.
This change removes the hibern8 state check for SCMI-based targets.

Change-Id: I5f681db5d63f1db57b3624e460e87f57d2874dba
Signed-off-by: Ram Kumar Dwivedi <rdwivedi@qti.qualcomm.com>