meta-qti-auto-kernel: Make I2C driver as modules in SA8797P
The GSI driver is loaded as a module and exports symbols used by
the I2C driver. However, commit 02a912ae8456 changed the I2C
driver from a module to built-in, which can lead to functionality
issues due to symbol dependency mismatch during boot.

Fix this by reverting the I2C driver to be built as a module in
the SA8797P (Nord-specific) configuration to ensure proper symbol
resolution and maintain expected functionality.

Change-Id: Ia94f9924e5614538c17e0a95428abe1505d167b8
Signed-off-by: Aniket Randive <arandive@qti.qualcomm.com>