meta-qti-auto-kernel: Ethernet: Add suspend and resume callbacks
Add suspend/resume platform operations, which, when populated, override
the init/exit platform operations when we suspend and resume. These
suspend()/resume() methods are called by core code, and thus are
designed to support any struct device, not just platform devices. This
allows them to be used by the PCI drivers we have.

Change-Id: Idf0ebac65cf540a00e4aa939327ee3a1d522a090